requirements.txt
requests>=2.24.0
slackclient>=2.9.3